Closer to desired build system
We got a lib folder that will have all the files for the MABDI Library.
We got a test and srs folder. Test has scrap ideas and src has
applications. But they both make executables. So let's concentrate on
the test folder (src works similar). In test/CMakeLists.txt it loops
through all sub folders and add a new target for each.